--- src/main/java/com/example/cinemax/repository/FilmRepository.java
@@ -1,4 +1,0 @@
-package com.example.cinemax.repository;
-
-public interface FilmRepository {
-}
