--- src/main/java/com/example/cinemax/repository/impl/FilmRepositoryImpl.java
@@ -1,9 +1,0 @@
-package com.example.cinemax.repository.impl;
-
-import com.example.cinemax.repository.FilmRepository;
-import com.example.cinemax.repository.UserRepository;
-import org.springframework.stereotype.Component;
-
-@Component
-public class FilmRepositoryImpl implements FilmRepository {
-}
