--- src/main/java/com/example/cinemax/repository/impl/SessionRepositoryImpl.java
@@ -1,9 +1,0 @@
-package com.example.cinemax.repository.impl;
-
-import com.example.cinemax.repository.FilmRepository;
-import com.example.cinemax.repository.SessionRepository;
-import org.springframework.stereotype.Component;
-
-@Component
-public class SessionRepositoryImpl implements SessionRepository {
-}
