--- src/main/java/com/example/cinemax/service/FilmService.java
@@ -1,5 +1,0 @@
-package com.example.cinemax.service;
-
-
-public interface FilmService {
-}
