+++ src/main/java/ua/edu/ukma/cinemax/controller/OAuthController.java
@@ -1,7 +1,7 @@
-package com.example.cinemax.controller;
+package ua.edu.ukma.cinemax.controller;
 
-import com.example.cinemax.repository.impl.UserCredentials;
-import com.example.cinemax.service.UserService;
+import ua.edu.ukma.cinemax.repository.impl.UserCredentials;
+import ua.edu.ukma.cinemax.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
