+++ src/main/java/ua/edu/ukma/cinemax/repository/FilmRepository.java
@@ -1,0 +1,4 @@
+package ua.edu.ukma.cinemax.repository;
+
+public interface FilmRepository {
+}
