+++ src/main/java/ua/edu/ukma/cinemax/repository/impl/FilmRepositoryImpl.java
@@ -1,0 +1,8 @@
+package ua.edu.ukma.cinemax.repository.impl;
+
+import ua.edu.ukma.cinemax.repository.FilmRepository;
+import org.springframework.stereotype.Component;
+
+@Component
+public class FilmRepositoryImpl implements FilmRepository {
+}
