+++ src/main/java/ua/edu/ukma/cinemax/repository/impl/SessionRepositoryImpl.java
@@ -1,0 +1,8 @@
+package ua.edu.ukma.cinemax.repository.impl;
+
+import ua.edu.ukma.cinemax.repository.SessionRepository;
+import org.springframework.stereotype.Component;
+
+@Component
+public class SessionRepositoryImpl implements SessionRepository {
+}
