+++ src/main/java/ua/edu/ukma/cinemax/service/FilmService.java
@@ -1,0 +1,5 @@
+package ua.edu.ukma.cinemax.service;
+
+
+public interface FilmService {
+}
