+++ src/main/java/ua/edu/ukma/cinemax/service/SessionService.java
@@ -1,4 +1,4 @@
-package com.example.cinemax.service;
+package ua.edu.ukma.cinemax.service;
 
 public interface SessionService {
 }
