+++ src/main/java/ua/edu/ukma/cinemax/service/UserService.java
@@ -1,0 +1,8 @@
+package ua.edu.ukma.cinemax.service;
+
+import ua.edu.ukma.cinemax.repository.impl.UserCredentials;
+import lombok.NonNull;
+
+public interface UserService {
+    UserCredentials getCredentials(@NonNull String userLogin);
+}
