+++ src/main/java/ua/edu/ukma/cinemax/service/impl/FilmServiceImpl.java
@@ -1,9 +1,7 @@
-package com.example.cinemax.service.impl;
+package ua.edu.ukma.cinemax.service.impl;
 
-import com.example.cinemax.repository.FilmRepository;
-import com.example.cinemax.repository.UserRepository;
-import com.example.cinemax.service.FilmService;
-import com.example.cinemax.service.UserService;
+import ua.edu.ukma.cinemax.repository.FilmRepository;
+import ua.edu.ukma.cinemax.service.FilmService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
