+++ src/test/java/ua/edu/ukma/cinemax/CinemaXApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.cinemax;
+package ua.edu.ukma.cinemax;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
